Refactor to improve go get workflow
This refactor is being done so that we can go get the repo and have it
clone the binary name as collins. It looked very ugly to have cmd and
cmds so cmds was moved to source and some bazel fixes were made as well.
