+++ src/assets/TestQueryCLIBasicWorkflow.json
