+++ src/assets/TestQueryCLIGetGPU.json
