+++ src/global.go
