+++ src/power.go
